requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv 
psycopg2